main: easy steps towards reducing heap usage
There's only going to ever be one ethernetTask/uart0Task run, so push these
large buffers off of the stack into .bss.  This moves 5KB into SRAM, but we
have enough slop to cope with it.

Start by reducing the amount of stack allocated for these tasks, 8192 words
is what ethernetTask had prior to these large buffers so let's just restore
that under the assumption it was bumped a lot and not necessarily
proportional to what was needed (4k / word size).

uart is reduced by 1024 *words* (not bytes), but I suspect even that is a
bit higher than needed.  We'll add a task to do some periodic stack sampling
shortly.